+++ api/src/routes/index.ts
@@ -4,7 +4,6 @@ import LoginRoutes from './login';
 import ProfileRoutes from './profiles';
 import review from './review'
 import TravelRoutes from './travel';
-import paymentCheckOut from './Stripe';
 import UploadDataFake from './uploadDataFake';
 
 import payments from './payments'
