+++ clientnative/Components/CompletarPerfil/CompleteProfileUser.js
@@ -29,8 +29,6 @@ const CompleteProfileUser = () => {
   }, [datosUser]);
 
 
-
-
   /// --> ESTADO PARA EL MODAL <-- ///
   const [isModalVisible, setisModalVisible] = useState(false);
   const [chooseData, setchooseData] = useState();
@@ -116,7 +114,6 @@ const CompleteProfileUser = () => {
     const obj = {
       identification: user.identification,
       zone: user.zone,
-      account: user.account,
       photo: selectedImage,
       id: datosUser.id
     };
@@ -155,7 +152,7 @@ const CompleteProfileUser = () => {
           <View style={styles.add}>
             <TouchableWithoutFeedback onPress={openImagePickerAsync}>
               <Image
-                source={require("../add-image.png")}
+                source={require("./add-image.png")}
                 style={styles.imgAdd}
               />
             </TouchableWithoutFeedback>
