--- clientnative/Components/helpers.js
@@ -1,15 +1,0 @@
-import { API_URL } from "./Config";
-import {Alert} from 'react-native'
-
-export async function fetchPublishableKey () {
-    try {
-        const response = await fetch(`${API_URL}/config`)
-        const { publishableKey } = await response.json()
-        return publishableKey
-    } catch (error) {
-        console.log(e)
-        console.warn('Unable to fetch publishable key. Is your server running?')
-        Alert.alert('Error', 'Unable to fetch publishable key. Is your server running?' );
-        return null
-    }
-};
