+++ clientnative/actions/index.js
@@ -2,10 +2,10 @@ import axios from "axios";
 export const GET_PRICE_QUOTE = "GET_PRICE_QUOTE";
 export const GET_TRAVELS = "GET_TRAVELS";
 export const POST_REQUEST_TRAVE = "POST_REQUEST_TRAVE";
-export const URL_PRICE_QUOTE = "http://192.168.1.6:3001/api/calculatePrice";
-export const URL_REQUEST_TRAVEL = "http://192.168.1.6:3001/api/requestTravel";
-export const URL_TRAVEL = "http://192.168.1.6:3001/api/Travel";
-export const URL_TRAVEL_ID = "http://192.168.1.6:3001/api/oneTravel";
+export const URL_PRICE_QUOTE = "http:/192.168.0.105:3001/api/calculatePrice";
+export const URL_REQUEST_TRAVEL = "http://192.168.0.105:3001/api/requestTravel";
+export const URL_TRAVEL = "http://192.168.0.105:3001/api/Travel";
+export const URL_TRAVEL_ID = "http://192.168.0.105:3001/api/oneTravel";
 export const GET_TRAVEL_ID = "GET_TRAVEL_ID";
 
 // import { registrarUsuario } from './index';
@@ -89,7 +89,7 @@ export function registrarUsuario(payload) {
   return async function (dispatch) {
     try {
       const response = await axios
-        .post("http://192.168.1.6:3001/api/user", payload) //aca cada uno pone su ip
+        .post("http://192.168.0.105:3001/api/user", payload) //aca cada uno pone su ip
         .then((r) => {
           dispatch({
             type: "REGISTROO",
@@ -109,7 +109,7 @@ export function logiarUsuario(payload) {
   return async function (dispatch) {
     try {
       const response = await axios
-        .post("http://192.168.1.6:3001/api/login", payload) //aca cada uno pone su ip
+        .post("http://192.168.0.105:3001/api/login", payload) //aca cada uno pone su ip
         .then((r) => {
           dispatch({
             type: "LOGEOO",
@@ -131,7 +131,7 @@ export function completeProfileUser(payload) {
   return async function (dispatch) {
     try {
       const response = await axios.post(
-        "http://192.168.1.6:3001/api/userProfile",
+        "http://192.168.0.105:3001/api/userProfile",
         payload
       );
       //console.log('Soy el console.log de response', response)
@@ -150,7 +150,7 @@ export function completeProfileCarrier(payload) {
   return async function (dispatch) {
     try {
       const response = await axios.post(
-        "http://192.168.1.6:3001/api/carrierProfile",
+        "http://192.168.0.105:3001/api/carrierProfile",
         payload
       );
       // console.log('Soy el console.log de response', response)
