+++ clientnative/package.json
@@ -14,6 +14,7 @@
     "@react-navigation/stack": "^6.0.11",
     "@stripe/stripe-react-native": "^0.2.3",
     "axios": "^0.24.0",
+    "dotenv": "^14.1.1",
     "expo": "~44.0.0",
     "expo-checkbox": "~2.0.0",
     "expo-image-picker": "^12.0.1",
